Improved the CreateUpdate forms
Also removed some files that are not used and renamed couple of components and variables for better reading.
