+++ src/app/admin/staff/page.tsx
@@ -1,13 +1,20 @@
-//here the member will be shown
+"use client";
 
 import { Box } from "@mui/material";
 import MemberCard from "src/components/MemberCard";
 import CreateUpdateMember from "src/components/CreateUpdateMember";
+import { useGetAllTreatments } from "src/core/react-query/treatments/hooks/useGetAllTreatmets";
 
 const AdminMemberPage = () => {
+  const { data: treatments, isLoading } = useGetAllTreatments();
+
+  if (isLoading) {
+    return "Loading...";
+  }
+
   return (
     <Box>
-      <CreateUpdateMember />
+      <CreateUpdateMember treatments={treatments} />
       <MemberCard />
     </Box>
   );
