+++ src/app/admin/treatments/[treatmentId]/page.tsx
@@ -1,14 +1,19 @@
 "use client";
 
-import { Box } from "@mui/material";
+import { useParams } from "next/navigation";
 import CreateUpdateTreatment from "src/components/CreateUpdateTreatment";
+import { useGetOneTreatment } from "src/core/react-query/treatments/hooks/useGetOneTreatment";
 
 const AdminUpdateTreatmentPage = () => {
-  return (
-    <Box>
-      <CreateUpdateTreatment />
-    </Box>
-  );
+  const params = useParams();
+  const treatmentId = String(params.treatmentId);
+  const { data: treatment, isLoading } = useGetOneTreatment({ treatmentId });
+
+  if (isLoading) {
+    return "Loading...";
+  }
+
+  return <CreateUpdateTreatment treatment={treatment} />;
 };
 
 export default AdminUpdateTreatmentPage;
