--- src/core/react-query/users/hooks/useGetAllUsers.ts
@@ -1,7 +1,0 @@
-import { useQuery } from "@tanstack/react-query";
-import { RequestTypes } from "../requestTypes";
-import { usersKeys } from "../usersKeys";
-
-export const useGetAllUsers = (payload: RequestTypes["getAllUsers"]) => {
-  return useQuery({ ...usersKeys.allUsers() });
-};
