--- src/core/react-query/users/index.ts
@@ -1,2 +1,0 @@
-export * from "./usersKeys";
-export * from "./requestTypes";
