--- src/core/react-query/users/types.ts
@@ -1,6 +1,0 @@
-export type User = {
-  userId: string;
-  username: string;
-  password: string;
-  confirmPassword?: string;
-};
